rename python scripts by its convention
